# internal/api/openapi.vohive.yaml
openapi: 3.0.3
info:
  title: VoHive Main API
  version: 1.0.0-main
  description: |
    VoHive 主服务控制面 API。
    - 服务器前缀：`/api`
    - 登录方式：`POST /auth/login`
    - 受保护接口统一使用 `Authorization: Bearer <token>`
    - 文档页面：`GET /docs`
servers:
  - url: /api
    description: VoHive main API prefix
security:
  - BearerAuth: []
...
paths:
...
  /command-center/commands:
    get:
      tags: [commands]
      summary: 获取共享命令目录
      responses:
        '200':
          description: 命令元数据
        '401':
          $ref: '#/components/responses/Unauthorized'
  /command-center/executions:
    post:
      tags: [commands]
      summary: 执行共享斜杠命令
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input]
              properties:
                input:
                  type: string
                  maxLength: 4096
      responses:
        '202':
          description: 命令已持久化并开始执行
        '400':
          $ref: '#/components/responses/ErrorMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /command-center/events:
    get:
      tags: [commands]
      summary: 按游标读取命令事件
      parameters:
        - $ref: '#/components/parameters/AfterEventID'
        - in: query
          name: before_id
          schema:
            type: integer
            format: int64
            minimum: 0
        - $ref: '#/components/parameters/EventLimit'
      responses:
        '200':
          description: 命令事件列表
        '401':
          $ref: '#/components/responses/Unauthorized'
  /command-center/stream:
    get:
      tags: [commands]
      summary: 命令事件 SSE 流
      description: 使用 Last-Event-ID 或 after_id 从持久化游标恢复，事件名为 command。
      parameters:
        - $ref: '#/components/parameters/AfterEventID'
        - in: header
          name: Last-Event-ID
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '200':
          description: SSE 命令事件流
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /command-center/history:
    delete:
      tags: [commands]
      summary: 清理已结束的命令历史
      responses:
        '200':
          description: 已删除的执行数量
        '401':
          $ref: '#/components/responses/Unauthorized'
  /command-center/recordings/{recording}:
    get:
      tags: [commands]
      summary: 读取命令产生的 VoWiFi 通话 MP3 录音
      parameters:
        - in: path
          name: recording
          required: true
          schema:
            type: string
            pattern: '^call_[A-Za-z0-9_.-]+\\.mp3$'
      responses:
        '200':
          description: MP3 通话录音
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/ErrorMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ErrorMessage'
  /balances:
    get:
      tags: [balance]
      summary: 获取余额查询历史
      parameters:
        - in: query
          name: device_id
          schema:
            type: string
        - in: query
          name: before
          schema:
            type: string
            format: date-time
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: 查询历史
        '401':
          $ref: '#/components/responses/Unauthorized'
  /devices/{device_id}/balance-queries:
    post:
      tags: [balance]
      summary: 为指定设备发起运营商余额查询
      parameters:
        - $ref: '#/components/parameters/DeviceIDPath'
      responses:
        '202':
          description: 查询已发起或已直接完成
        '400':
          $ref: '#/components/responses/ErrorMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/ErrorMessage'
    get:
      tags: [balance]
      summary: 获取指定设备的余额查询历史
      parameters:
        - $ref: '#/components/parameters/DeviceIDPath'
        - in: query
          name: before
          schema:
            type: string
            format: date-time
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: 查询历史
        '401':
          $ref: '#/components/responses/Unauthorized'
  /carrier-query-rules:
    get:
      tags: [balance]
      summary: 获取内置及自定义运营商余额规则
      responses:
        '200':
          description: 内置规则和自定义覆盖
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [balance]
      summary: 创建自定义运营商余额规则
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarrierQueryRule'
      responses:
        '200':
          description: 已保存规则
        '400':
          $ref: '#/components/responses/ErrorMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /carrier-query-rules/{rule_id}:
    put:
      tags: [balance]
      summary: 更新自定义运营商余额规则
      parameters:
        - in: path
          name: rule_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarrierQueryRule'
      responses:
        '200':
          description: 已保存规则
        '400':
          $ref: '#/components/responses/ErrorMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [balance]
      summary: 删除自定义运营商余额规则
      parameters:
        - in: path
          name: rule_id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 已删除
        '401':
          $ref: '#/components/responses/Unauthorized'
  /commands/catalog:
    get:
      tags: [commands]
      summary: 获取共享命令目录
      responses:
        '200':
          description: 命令元数据
          content:
            application/json:
              schema:
                type: object
                properties:
                  commands:
                    type: array
                    items:
                      $ref: '#/components/schemas/CommandDefinition'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /commands/executions:
    post:
      tags: [commands]
      summary: 执行共享斜杠命令
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input]
              properties:
                input:
                  type: string
                  maxLength: 4096
      responses:
        '202':
          description: 命令已持久化并开始执行
          content:
            application/json:
              schema:
                type: object
                properties:
                  execution:
                    $ref: '#/components/schemas/CommandExecution'
        '400':
          $ref: '#/components/responses/ErrorMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /commands/events:
    get:
      tags: [commands]
      summary: 按游标读取命令事件
      parameters:
        - $ref: '#/components/parameters/AfterEventID'
        - in: query
          name: before_id
          description: 向前读取此事件 ID 之前的历史；传 0 读取最新一页。
          schema:
            type: integer
            format: int64
            minimum: 0
        - $ref: '#/components/parameters/EventLimit'
      responses:
        '200':
          description: 命令事件列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/CommandEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /commands/events/stream:
    get:
      tags: [commands]
      summary: 命令事件 SSE 流
      description: 使用 Last-Event-ID 或 after_id 从持久化游标恢复，事件名为 command。
      parameters:
        - $ref: '#/components/parameters/AfterEventID'
        - in: header
          name: Last-Event-ID
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '200':
          description: SSE 命令事件流
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /commands/history:
    delete:
      tags: [commands]
      summary: 清理已结束的命令历史
      description: 只删除 completed 和 failed，保留 running 执行。
      responses:
        '200':
          description: 已删除的执行数量
        '401':
          $ref: '#/components/responses/Unauthorized'
  /balance/queries:
    post:
      tags: [balance]
      summary: 手动发起运营商余额查询
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_id]
              properties:
                device_id:
                  type: string
      responses:
        '202':
          description: 查询已发起或已直接完成
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    $ref: '#/components/schemas/BalanceQuery'
        '400':
          $ref: '#/components/responses/ErrorMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/ErrorMessage'
    get:
      tags: [balance]
      summary: 获取余额查询历史
      parameters:
        - in: query
          name: device_id
          schema:
            type: string
        - in: query
          name: before
          schema:
            type: string
            format: date-time
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: 查询历史
        '401':
          $ref: '#/components/responses/Unauthorized'
  /balance/queries/{query_id}:
    get:
      tags: [balance]
      summary: 获取单次余额查询
      parameters:
        - in: path
          name: query_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 查询详情
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ErrorMessage'
  /balance/rules:
    get:
      tags: [balance]
      summary: 获取内置及自定义运营商余额规则
      responses:
        '200':
          description: 内置规则和自定义覆盖
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /balance/rules/{rule_id}:
    put:
      tags: [balance]
      summary: 新增或更新自定义余额规则
      parameters:
        - in: path
          name: rule_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarrierQueryRule'
      responses:
        '200':
          description: 已保存规则
        '400':
          $ref: '#/components/responses/ErrorMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [balance]
      summary: 删除自定义余额规则
      parameters:
        - in: path
          name: rule_id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 已删除
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SessionToken
  parameters:
    DeviceIDPath:
      in: path
      name: device_id
      required: true
      schema:
        type: string
      description: 设备 ID
...
    AfterEventID:
      in: query
      name: after_id
      schema:
        type: integer
        format: int64
        minimum: 0
        default: 0
    EventLimit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 100
  responses:
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ErrorMessage:
      description: 错误信息
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        error:
          type: string
        warning:
          type: string
        result:
          type: object
          additionalProperties: true
        busy:
          type: boolean
    CommandDefinition:
      type: object
      required: [name, usage, summary, dangerous, async, device_argument]
      properties:
        name:
          type: string
        usage:
          type: string
        summary:
          type: string
        dangerous:
          type: boolean
        async:
          type: boolean
        device_argument:
          type: boolean
    CommandExecution:
      type: object
      required: [id, input, command, state, created_at, updated_at]
      properties:
        id:
          type: string
        input:
          type: string
        command:
          type: string
        state:
          type: string
          enum: [running, completed, failed]
        error:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CommandEvent:
      type: object
      required: [id, execution_id, kind, text, created_at]
      properties:
        id:
          type: integer
          format: int64
        execution_id:
          type: string
        kind:
          type: string
          enum: [accepted, progress, result, error]
        text:
          type: string
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/CommandAttachment'
        execution:
          $ref: '#/components/schemas/CommandExecution'
        created_at:
          type: string
          format: date-time
    CommandAttachment:
      type: object
      required: [type, recording, content_type]
      properties:
        type:
          type: string
          enum: [audio]
        recording:
          type: string
        content_type:
          type: string
          enum: [audio/mpeg]
    BalanceQuery:
      type: object
      required: [id, device_id, iccid, rule_id, transport, state, parse_state, started_at, expires_at]
      properties:
        id:
          type: string
        device_id:
          type: string
        iccid:
          type: string
        rule_id:
          type: string
        transport:
          type: string
          enum: [sms, ussd]
        state:
          type: string
          enum: [sending, awaiting_reply, completed, timed_out, failed]
        parse_state:
          type: string
          enum: [pending, parsed, unparsed]
        amount:
          type: string
        currency:
          type: string
        summary:
          type: string
        raw_response:
          type: string
        error:
          type: string
    CarrierQueryRule:
      type: object
      required: [mcc, mnc, operator, transport, response_mode, enabled]
      properties:
        id:
          type: string
        mcc:
          type: string
        mnc:
          type: string
        spn:
          type: string
        variant:
          type: string
        operator:
          type: string
        transport:
          type: string
          enum: [sms, ussd, unsupported]
        destination:
          type: string
        payload:
          type: string
        response_mode:
          type: string
          enum: [direct, sms, none]
        expected_senders:
          type: array
          items:
            type: string
        parser_pattern:
          type: string
        currency:
          type: string
        cost_status:
          type: string
        evidence_type:
          type: string
        evidence_url:
          type: string
          format: uri
        limitations:
          type: array
          items:
            type: string
        alternative:
          type: string
        enabled:
          type: boolean
...